refactor: move `Statuses` to components dir
